implementing deposit unit test helpers
